Move Capacity class to model/booking
