+++ src/main/java/seedu/address/logic/commands/UpdateCapacityCommand.java
@@ -4,8 +4,8 @@ import static java.util.Objects.requireNonNull;
 
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.Capacity;
 import seedu.address.model.Model;
+import seedu.address.model.booking.Capacity;
 
 /**
  * Updates the capacity of the restaurant.
