+++ src/main/java/seedu/address/model/BookingModel.java
@@ -2,6 +2,7 @@ package seedu.address.model;
 
 import javafx.collections.ObservableList;
 import seedu.address.model.booking.Booking;
+import seedu.address.model.booking.Capacity;
 
 /**
  * The API that stores the booking side of the model.
